server/requirements.txt
# Production dependencies
Flask==3.0.2
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
email-validator==2.1.0.post1
python-json-logger==2.0.7
uuid==1.30
gunicorn==21.2.0
psycopg2-binary==2.9.9  # For PostgreSQL support
Flask-SQLAlchemy==3.1.1  # For database ORM
Flask-Migrate==4.0.5  # For database migrations

# Testing dependencies
pytest==8.0.2
pytest-cov==4.1.0 